Update: move row_encoder from kernel to row
