+++ cpp/src/arrow/compute/CMakeLists.txt
@@ -93,6 +93,7 @@ add_arrow_test(internals_test
                row/compare_test.cc
                row/grouper_test.cc
                row/row_test.cc
+               row/row_encoder_internal_test.cc
                util_internal_test.cc)
 
 add_arrow_compute_test(expression_test SOURCES expression_test.cc)
