+++ cpp/src/arrow/compute/kernels/CMakeLists.txt
@@ -137,5 +137,4 @@ add_arrow_compute_test(aggregate_test
 # ----------------------------------------------------------------------
 # Utilities
 
-add_arrow_compute_test(kernel_utility_test SOURCES codegen_internal_test.cc
-                       row_encoder_internal_test.cc)
+add_arrow_compute_test(kernel_utility_test SOURCES codegen_internal_test.cc)
